PoC for TDVF TD_VMM_DATA
This patch implements a PoC introducing a new type of resource that can
be declared through the TD_HOB section. The point is to let the VMM
share some data with the guest through one or multiple TD_VMM_DATA
regions.

This will be particularly useful to share the ACPI tables and SMBIOS
table for VMMs with no fw_cfg support.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>